feat: move completion_rate calculation to client side
