+++ lib/backend/addToChildren.ts
@@ -1,6 +1,5 @@
 import { getDueDate } from '../parser';
 import { GithubIssueDataWithGroupAndChildren, IssueData } from '../types';
-import { calculateCompletionRate } from './calculateCompletionRate';
 import { ErrorManager } from './errorManager';
 
 export function addToChildren(
@@ -22,12 +21,12 @@ export function addToChildren(
       html_url: parent.html_url,
       labels: parent.labels,
       node_id: parent.node_id,
-      completion_rate: calculateCompletionRate(parent),
+      completion_rate: 0, // calculated on the client-side once all issues are loaded
       due_date: parentDueDate,
     };
     return data.map((item: GithubIssueDataWithGroupAndChildren): IssueData => ({
       labels: item.labels ?? [],
-      completion_rate: calculateCompletionRate(item),
+      completion_rate: 0, // calculated on the client-side once all issues are loaded
       due_date: getDueDate(item, errorManager).eta,
       html_url: item.html_url,
       group: item.group,
