+++ lib/calculateCompletionRate.ts
@@ -1,5 +1,6 @@
-import { IssueStates } from '../enums';
-import { IssueData } from '../types';
+import { State } from '@hookstate/core';
+import { IssueStates } from './enums';
+import { IssueData } from './types';
 
 export type CalculateCompletionRateOptions = Pick<IssueData, 'html_url' | 'state'> & { children: CalculateCompletionRateOptions[] };
 
@@ -28,6 +29,7 @@ function getIssueCounts(issueStatesMap: Map<string, IssueStates>): GetIssueCount
   let open = 0;
   let closed = 0;
   issueStatesMap.forEach((value) => {
+    console.log(`value: `, value);
     if (value === IssueStates.OPEN) {
       open++;
     } else {
@@ -45,6 +47,20 @@ function getIssueCounts(issueStatesMap: Map<string, IssueStates>): GetIssueCount
 
 export function calculateCompletionRate (issue: CalculateCompletionRateOptions): number {
   const issueStatesMap = getIssueStatesMap(issue);
-  const { percentClosed } = getIssueCounts(issueStatesMap);
+  console.log(`issue.html_url: `, issue.html_url);
+  const issueCounts = getIssueCounts(issueStatesMap);
+  console.log(`issueCounts: `, issueCounts);
+  const { percentClosed } = issueCounts;
   return percentClosed;
 };
+
+export function assignCompletionRateToIssues (issue: State<IssueData> | State<IssueData | null>): void {
+  if (issue.ornull === null) {
+    return
+  }
+  const completion_rate = calculateCompletionRate({ html_url: issue.ornull.html_url.value, state: issue.ornull.state.value, children: issue.ornull.children.value });
+  issue.merge((issue) => ({ ...issue, completion_rate }))// = completionRate
+  issue.ornull.children.forEach((child) => {
+    assignCompletionRateToIssues(child);
+  })
+}
